Update Pipfile.lock
In it's original state this won't compile cffi on Python3.  Older versions of cffi don't seem to have any error handling in compilation for PyInterpreterState.  I've updated the Pipfile.lock with the working version and correct SHA256 which allows pipenv sync to download and build all dependencies, and ultimately run.